==> jupyter_notebooks/05_predict_cancellation.ipynb <==
# **Predict cancellation (classification) notebook**

## Objectives

* Fit and evaluate a classification model to predict if a booking will be cancelled or not.

## Inputs

* outputs/datasets/collection/hotel_bookings.csv
* Instructions on which variables to use for data cleaning and feature engineering. They are found in each respective notebook.

## Outputs

* Train set (features and target)
* Test set (features and target)
* Data cleaning and Feature Engineering pipeline
* Modeling pipeline
* Feature importance plot


---

# Change working directory

Change the working directory from its current folder to its parent folder
* Access the current directory with os.getcwd()

import os
current_dir = os.getcwd()
current_dir

Make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

---

# Load data

import numpy as np
import pandas as pd

# Drop variables not needed for this project
df = (
    pd.read_csv("outputs/datasets/collection/hotel_bookings.csv")
    .drop(
        labels=[
            'company',
            'agent',
            'country',
            'arrival_date_year',
            'reservation_status',
            'reservation_status_date',
            'assigned_room_type'
        ],
        axis=1
    )  
)

print(df.shape)
df.head(3)

# ML pipeline with all data

## ML pipeline for Data Cleaning and Feature Engineering

from sklearn.pipeline import Pipeline

# Feature Engineering
from feature_engine.selection import SmartCorrelatedSelection
from feature_engine.encoding import OrdinalEncoder
from feature_engine.imputation import ArbitraryNumberImputer


def PipelineDataCleaningAndFeatureEngineering():
    pipeline_base = Pipeline(
        [
            (
                "OrdinalEncoder",
                OrdinalEncoder(
                    encoding_method='arbitrary',
                    variables=[
                        'hotel',
                        'arrival_date_month',
                        'meal',
                        'market_segment',
                        'distribution_channel',
                        'reserved_room_type',
                        'deposit_type',
                        'customer_type'
                    ]
                )
            ),

            (
                "SmartCorrelatedSelection",
                SmartCorrelatedSelection(
                    variables=None,
                    method="spearman",
                    threshold=0.6,
                    selection_method="variance"
                )
            )
        ]
    )

    return pipeline_base


PipelineDataCleaningAndFeatureEngineering()

## ML pipeline for modelling and hyperparameter optimisation

# Feat Scaling
from sklearn.preprocessing import StandardScaler

# Feat Selection
from sklearn.feature_selection import SelectFromModel

# ML algorithms
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.ensemble import ExtraTreesClassifier
from sklearn.ensemble import AdaBoostClassifier
from xgboost import XGBClassifier


def PipelineClf(model):
    pipeline_base = Pipeline([
        ("scaler", StandardScaler()),
        ("feat_selection", SelectFromModel(model)),
        ("model", model),
    ])

    return pipeline_base

Custom Class for Hyperparameter Optimisation.

from sklearn.model_selection import GridSearchCV


class HyperparameterOptimizationSearch:

    def __init__(self, models, params):
        self.models = models
        self.params = params
        self.keys = models.keys()
        self.grid_searches = {}

    def fit(self, X, y, cv, n_jobs, verbose=1, scoring=None, refit=False):
        for key in self.keys:
            print(f"\nRunning GridSearchCV for {key} \n")

            model = PipelineClf(self.models[key])
            params = self.params[key]
            gs = GridSearchCV(model, params, cv=cv, n_jobs=n_jobs,
                            verbose=verbose, scoring=scoring, )
            gs.fit(X, y)
            self.grid_searches[key] = gs

    def score_summary(self, sort_by='mean_score'):
        def row(key, scores, params):
            d = {
                'estimator': key,
                'min_score': min(scores),
                'max_score': max(scores),
                'mean_score': np.mean(scores),
                'std_score': np.std(scores),
            }
            return pd.Series({**params, **d})

        rows = []
        for k in self.grid_searches:
            params = self.grid_searches[k].cv_results_['params']
            scores = []
            for i in range(self.grid_searches[k].cv):
                key = "split{}_test_score".format(i)
                r = self.grid_searches[k].cv_results_[key]
                scores.append(r.reshape(len(params), 1))

            all_scores = np.hstack(scores)
            for p, s in zip(params, all_scores):
                rows.append((row(k, s, p)))

        df = pd.concat(rows, axis=1).T.sort_values([sort_by], ascending=False)
        columns = ['estimator', 'min_score',
                'mean_score', 'max_score', 'std_score']
        columns = columns + [c for c in df.columns if c not in columns]
        return df[columns], self.grid_searches

## Split Train and Test set

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(
    df.drop(['is_canceled'], axis=1),
    df['is_canceled'],
    test_size=0.2,
    random_state=0,
)

print(X_train.shape, y_train.shape, X_test.shape, y_test.shape)

## Handle target imbalance

pipeline_data_cleaning_feat_eng = PipelineDataCleaningAndFeatureEngineering()
X_train = pipeline_data_cleaning_feat_eng.fit_transform(X_train)
X_test = pipeline_data_cleaning_feat_eng.transform(X_test)
print(X_train.shape, y_train.shape, X_test.shape, y_test.shape)

Check Train set target distribution.

import matplotlib.pyplot as plt
import seaborn as sns
sns.set_style("whitegrid")

y_train.value_counts().plot(kind='bar', title='Train Set Target Distribution')
plt.show()

Use SMOTE (Synthetic Minority Oversampling TEchnique) to balance Train set target.

from imblearn.over_sampling import SMOTE

oversample = SMOTE(sampling_strategy='minority', random_state=0)
X_train, y_train = oversample.fit_resample(X_train, y_train)
print(X_train.shape, y_train.shape, X_test.shape, y_test.shape)

Check Train set target distribution after resampling.

import matplotlib.pyplot as plt

y_train.value_counts().plot(kind='bar', title='Train Set Target Distribution')
plt.show()

# Grid Search CV - Sklearn

## Use standard hyperparameters to find most suitable algorithm

models_quick_search = {
    "LogisticRegression": LogisticRegression(random_state=0),
    "XGBClassifier": XGBClassifier(random_state=0),
    "DecisionTreeClassifier": DecisionTreeClassifier(random_state=0),
    "RandomForestClassifier": RandomForestClassifier(random_state=0),
    "GradientBoostingClassifier": GradientBoostingClassifier(random_state=0),
    "ExtraTreesClassifier": ExtraTreesClassifier(random_state=0),
    "AdaBoostClassifier": AdaBoostClassifier(random_state=0),
}

params_quick_search = {
    "LogisticRegression": {},
    "XGBClassifier": {},
    "DecisionTreeClassifier": {},
    "RandomForestClassifier": {},
    "GradientBoostingClassifier": {},
    "ExtraTreesClassifier": {},
    "AdaBoostClassifier": {},
}

Quick GridSearch CV - Binary Classifier

from sklearn.metrics import make_scorer, recall_score

search = HyperparameterOptimizationSearch(
    models=models_quick_search,
    params=params_quick_search
)
search.fit(
    X_train,
    y_train,
    scoring=make_scorer(recall_score, pos_label=1),
    n_jobs=-1,
    cv=5
)

Check results.

grid_search_summary, grid_search_pipelines = search.score_summary(sort_by='mean_score')
grid_search_summary

## Do an extensive search on the most suitable algorithm to find the best hyperparameter configuration

Define model and parameters for Extensive Search.

models_search = {
    "ExtraTreesClassifier": ExtraTreesClassifier(random_state=0)
    # "RandomForestClassifier": RandomForestClassifier(random_state=0)
    # "DecisionTreeClassifier": DecisionTreeClassifier(random_state=0)
    # "GradientBoostingClassifier": GradientBoostingClassifier(random_state=0),
    # "AdaBoostClassifier": AdaBoostClassifier(random_state=0),
    # "XGBClassifier": XGBClassifier(random_state=0),
    # "LogisticRegression": LogisticRegression(random_state=0),
}

params_search = {
    "ExtraTreesClassifier": {
        'model__n_estimators': [140],
        'model__max_depth': [None],
        'model__min_samples_split': [3],
        'model__min_samples_leaf': [2]
    }
    # "RandomForestClassifier": {
    #     'model__n_estimators': [140],
    #     'model__max_depth': [None],
    #     'model__min_samples_split': [2],
    #     'model__min_samples_leaf': [1],
    #     'model__max_leaf_nodes': [50],
    # }
    # "DecisionTreeClassifier": {
    #     'model__max_depth': [4],
    #     'model__min_samples_split': [2],
    #     'model__min_samples_leaf': [1],
    #     'model__max_leaf_nodes': [None]
    # }
    # "XGBClassifier": {
    #     'model__n_estimators': [30,80,200],
    #     'model__max_depth': [None, 3, 15],
    #     'model__learning_rate': [0.01,0.1,0.001],
    #     'model__gamma': [0, 0.1]
    # }
}

Extensive GridSearch CV - Binary Classifier

from sklearn.metrics import recall_score, make_scorer

search = HyperparameterOptimizationSearch(
    models=models_search,
    params=params_search
)
search.fit(
    X_train,
    y_train,
    scoring=make_scorer(recall_score, pos_label=1),
    n_jobs=-1,
    cv=5
)

Check results.

grid_search_summary, grid_search_pipelines = search.score_summary(sort_by='mean_score')
grid_search_summary

Get best model name programmatically.

best_model = grid_search_summary.iloc[0,0]
best_model

Parameters for best model

best_parameters = grid_search_pipelines[best_model].best_params_
best_parameters

Define the best clf pipeline.

pipeline_clf = grid_search_pipelines[best_model].best_estimator_
pipeline_clf

## Assess feature importance

X_train.head(3)

With the current model, assess with `.features_importances_`.

# Create DataFrame to display feature importance
df_feature_importance = (pd.DataFrame(data={
    'Feature': X_train.columns[pipeline_clf['feat_selection'].get_support()],
    'Importance': pipeline_clf['model'].feature_importances_})
    .sort_values(by='Importance', ascending=False)
)

# Re-assign best_features order
best_features = df_feature_importance['Feature'].to_list()

# Most important features statement and plot
print(f"* These are the {len(best_features)} most important features in descending order. "
    f"The model was trained on them: \n{df_feature_importance['Feature'].to_list()}")

df_feature_importance.plot(kind='bar', x='Feature', y='Importance')
plt.show()

## Evaluate pipeline on Train and Test sets

from sklearn.metrics import classification_report, confusion_matrix


def confusion_matrix_and_report(X, y, pipeline, label_map):

    prediction = pipeline.predict(X)

    print('---  Confusion Matrix  ---')
    print(pd.DataFrame(confusion_matrix(y_true=prediction, y_pred=y),
        columns=[["Actual " + sub for sub in label_map]],
        index=[["Prediction " + sub for sub in label_map]]
        ))
    print("\n")

    print('---  Classification Report  ---')
    print(classification_report(y, prediction, target_names=label_map), "\n")


def clf_performance(X_train, y_train, X_test, y_test, pipeline, label_map):
    print("#### Train Set #### \n")
    confusion_matrix_and_report(X_train, y_train, pipeline, label_map)

    print("#### Test Set ####\n")
    confusion_matrix_and_report(X_test, y_test, pipeline, label_map)

Evaluation: cross check with metrics defined at ML business case.

- 80% Recall for cancellation on train and test set.

clf_performance(
    X_train=X_train,
    y_train=y_train,
    X_test=X_test,
    y_test=y_test,
    pipeline=pipeline_clf,
    label_map= ['not cancelled', 'cancelled']
)

## Refit pipeline with best features

### Rewrite ML pipeline for Data Cleaning and Feature Engineering

best_features

New Pipeline for DataCleaning And FeatureEngineering.

def PipelineDataCleaningAndFeatureEngineering():
    pipeline_base = Pipeline(
        [
            (
                "OrdinalEncoder",
                OrdinalEncoder(
                    encoding_method='arbitrary',
                    variables=[
                        'market_segment',
                        'deposit_type'
                    ]
                )
            )
        ]
    )

    return pipeline_base

### Rewrite ML pipeline for Data Cleaning and Feature Engineering

Function for Pipeline optmisation.

# Pipeline Optmization: Model
def PipelineClf(model):
    pipeline_base = Pipeline([
        ("scaler", StandardScaler()),
        ("model", model)
    ])

    return pipeline_base

### Split Train Test Set, considering only with best features

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(
    df.drop(['is_canceled'], axis=1),
    df['is_canceled'],
    test_size=0.2,
    random_state=0,
)

print(X_train.shape, y_train.shape, X_test.shape, y_test.shape)

Filter only the most important variables.

X_train = X_train.filter(best_features)
X_test = X_test.filter(best_features)

print(X_train.shape, y_train.shape, X_test.shape, y_test.shape)
X_train.head(3)

### Handle target imbalance

pipeline_data_cleaning_feat_eng = PipelineDataCleaningAndFeatureEngineering()
X_train = pipeline_data_cleaning_feat_eng.fit_transform(X_train)
X_test = pipeline_data_cleaning_feat_eng.transform(X_test)
print(X_train.shape, y_train.shape, X_test.shape, y_test.shape)

Check Train Set Target distribution

import matplotlib.pyplot as plt

y_train.value_counts().plot(kind='bar', title='Train Set Target Distribution')
plt.show()

Use SMOTE to balance Train Set target.

from imblearn.over_sampling import SMOTE

oversample = SMOTE(sampling_strategy='minority', random_state=0)
X_train, y_train = oversample.fit_resample(X_train, y_train)
print(X_train.shape, y_train.shape, X_test.shape, y_test.shape)

Check Train Set Target distribution after SMOTE.

y_train.value_counts().plot(kind='bar',title='Train Set Target Distribution')
plt.show()

## Grid Search CV: Sklearn

Using the most suitable model from the last section and its best hyperparameter configuration.

Use the same model from the last GridCV search.

models_search

Use the best parameters from the last GridCV search.

best_parameters

Type in manually since the hyperparameter values have to be a list. The previous dictionary is not in this format.

params_search = {
    'ExtraTreesClassifier': {
        'model__max_depth': [None],
        'model__min_samples_leaf': [2],
        'model__min_samples_split': [3],
        'model__n_estimators': [140]
    }
}
params_search

GridSearch CV

from sklearn.metrics import recall_score, make_scorer

quick_search = HyperparameterOptimizationSearch(
    models=models_search, params=params_search)
quick_search.fit(
    X_train,
    y_train,
    scoring=make_scorer(recall_score, pos_label=1),
    n_jobs=-1,
    cv=5
)

Check results.

grid_search_summary, grid_search_pipelines = quick_search.score_summary(sort_by='mean_score')
grid_search_summary 

Define the best clf pipeline.

best_model = grid_search_summary.iloc[0, 0]
pipeline_clf = grid_search_pipelines[best_model].best_estimator_
pipeline_clf

## Assess feature importance

best_features = X_train.columns

# create DataFrame to display feature importance
df_feature_importance = (pd.DataFrame(data={
    'Feature': best_features,
    'Importance': pipeline_clf['model'].feature_importances_})
    .sort_values(by='Importance', ascending=False)
)


# Most important features statement and plot
print(f"* These are the {len(best_features)} most important features in descending order. "
    f"The model was trained on them: \n{df_feature_importance['Feature'].to_list()}")

df_feature_importance.plot(kind='bar', x='Feature', y='Importance')
plt.show()

## Evaluate Pipeline on Train and Test Sets

Cross-check with metrics defined in the ML business case.

- 80% Recall for cancellation on train and test set.

clf_performance(
    X_train=X_train,
    y_train=y_train,
    X_test=X_test,
    y_test=y_test,
    pipeline=pipeline_clf,
    label_map= ['not cancelled', 'cancelled']
)

The above performance means that the model is overfitting as there is a considerable difference. Future improvements will be made to fine tune model performance.

---

# Push files to Repo

Generate the following files:

- Train set
- Test set
- Data cleaning and Feature Engineering pipeline
- Modeling pipeline
- Features importance plot

import joblib
import os

version = 'v1'
file_path = f'outputs/ml_pipeline/predict_cancellation/{version}'

try:
    os.makedirs(name=file_path)
except Exception as e:
    print(e)

### Train Set

Note that the variables are transformed already in X_train and the shape is 7884 - after SMOTE was applied.

print(X_train.shape)
X_train.head()

X_train.to_csv(f"{file_path}/X_train.csv", index=False)

y_train

y_train.to_csv(f"{file_path}/y_train.csv", index=False)

### Test Set

Note that the variables are transformed already in X_test.

print(X_test.shape)
X_test.head()

X_test.to_csv(f"{file_path}/X_test.csv", index=False)

y_test

y_test.to_csv(f"{file_path}/y_test.csv", index=False)

### ML Pipelines: Data Cleaning and Feat Eng pipeline and Modelling Pipeline

Save 2 pipelines:

- Both should be used in conjunction to predict Live Data.
- To predict on Train Set, Test Set we use only pipeline_clf, since the data is already processed.

Pipeline responsible for Data Cleaning and Feature Engineering.

pipeline_data_cleaning_feat_eng

joblib.dump(
    value=pipeline_data_cleaning_feat_eng,
    filename=f"{file_path}/clf_pipeline_data_cleaning_feat_eng.pkl"
)

Pipeline responsible for Feature Scaling, and Model.

pipeline_clf

joblib.dump(
    value=pipeline_clf,
    filename=f"{file_path}/clf_pipeline_model.pkl"
)

### Feature Importance plot

df_feature_importance.plot(kind='bar',x='Feature',y='Importance')
plt.show()

df_feature_importance.plot(kind='bar', x='Feature', y='Importance')
plt.savefig(f'{file_path}/features_importance.png', bbox_inches='tight')

---

# Next steps

Create dashboard.
